entra-security-agent/requirements.txt
openai>=1.40.0
azure-identity>=1.17.0
msgraph-sdk>=1.5.0
python-dotenv>=1.0.0
